Keep all search views in the search app
